Fix value type to be saved in genesis block

+++ governance/default.go
@@ -1134,7 +1134,7 @@ func GetGovernanceItemsFromChainConfig(config *params.ChainConfig) GovernanceSet
 	if config.IsKoreForkEnabled(common.Big0) &&
 		config.Governance != nil {
 		governanceMap := map[int]interface{}{
-			params.DeriveShaImpl: config.DeriveShaImpl,
+			params.DeriveShaImpl: uint64(config.DeriveShaImpl), // checkValueType expects uint64
 		}
 		if !common.EmptyAddress(config.Governance.GovParamContract) {
 			governanceMap[params.GovParamContract] = config.Governance.GovParamContract
